requirements.txt
Django==2.2.3
djangorestframework==3.10.0
djangorestframework-simplejwt==4.3.0